refactor: remove ToastInjector in favor of creating an instance via Injector.create
This commit removes a custom ToastInjector (that provided a single token) in favor of creating a new injector instance using the `Injector.create` call.

The goal of this change is to make the library forward-compatible with an upcoming change to the Injector interface, see https://github.com/angular/angular/pull/46761.